Reduce arc usages; add split method

+++ src/util.rs
@@ -1,5 +1,5 @@
 pub(crate) fn divide_ternary_sizes(size: usize) -> (usize, usize, usize) {
-  let group_size = (size / 3) as usize;
+  let group_size = size / 3;
   let extra = size - group_size * 3;
 
   (group_size, group_size + extra, group_size)
